[fix][VCF,indel] Add check for deletin in alleles
Bug: VCF v4.2 introduced the use of asterisk notation (*) in ALTs to
represent spanning deletions. If there are alternative alleles at a
site, the ‘*’ allele is reserved to indicate that an allele allele is
missing due to an upstream deletion. This led to segfaults in angsd VCF
reading.
Fix: Add hasDelAlt func to check if there is any deletion * in alleles
using strcmp.

Fixes #557