[draftFeat] Add doFasta with vcf files
Add bcfReader and -bcf option, add required header bitsetting, read
fmt/ad tag and fill funkyPars counts array to be passed to doFasta. This
makes it possible to use vcf input with -doFasta 1 and 2 to call
pseudohaplo fastas from vcf files. It is a quick workaround for a
functionality I needed. Cleaning of the counts array is left to
abcWriteFasta cleaner. Ideally this should be done in doCounts and
cleaned in doCounts cleaner like bam input, but this was quicker since
we need nInd and nAlleles info to initialize counts array sizes: there
is currently no direct interaction between vcf reading and doCounts after
reading vcf header but before reading bcf records. Therefore I am
pushing this as a draft feat for version control but not for directly
adding to main program.